requirements.txt
streamlit
typing-extensions
langchain
langchain-deepseek
openai
langgraph
tavily-python